Attendance system/requirements.txt
Flask
requests
opencv-python
face-recognition
Pillow
mtcnn
numpy==1.26.4
